+++ src/main/java/com/packt/web/Dao/RoleDaoInterface.java
@@ -4,6 +4,6 @@ import com.packt.web.bean.Role;
 
 public interface RoleDaoInterface {
     
-    public Role getRole(int id);
+    public Role getRole(Long id);
  
 }
