+++ src/main/java/com/packt/web/Dao/UserRatingDao.java
@@ -4,14 +4,12 @@ import java.math.BigDecimal;
 import java.util.List;
 
 import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.packt.web.bean.User;
 import com.packt.web.bean.UserRating;
 
 @Repository
@@ -33,13 +31,13 @@ public class UserRatingDao implements UserRatingDaoInterface{
 	}
 	
 	@Transactional
-	public void removeUserRating(int id) {//trip.id
+	public void removeUserRating(Long id) {//trip.id
 		Query query = entityManager.createQuery("delete UserRating where trip.id="+id+"");
 		query.executeUpdate();
 	}	
 
 	@Transactional
-	public String getUserRating(int userId) {
+	public String getUserRating(Long userId) {
 	   
 		Query query = entityManager.createNativeQuery("SELECT AVG(rating) FROM `userrating` WHERE user_id="+userId+"");
         BigDecimal result= (BigDecimal) query.getSingleResult();
@@ -51,7 +49,7 @@ public class UserRatingDao implements UserRatingDaoInterface{
 	}
 	
 	@Transactional
-	public boolean isRatedByUser(int tripId, int authorId) {
+	public boolean isRatedByUser(Long tripId, Long authorId) {
 	  
 		Query query = entityManager.createQuery("select r from UserRating r where r.author.id="+authorId+" and r.trip.id="+tripId+"");
 		List<UserRating> result = query.getResultList();
