+++ src/main/java/com/packt/web/service/MessageService.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.packt.web.Dao.MessageDaoInterface;
-import com.packt.web.Dao.UserDaoInterface;
 import com.packt.web.bean.Message;
 import com.packt.web.bean.User;
 
@@ -20,17 +19,17 @@ public class MessageService implements MessageServiceInterface{
 	@Autowired
 	UserServiceInterface userService;
 	
-	public List<Message> getMessage(int id) {
+	public List<Message> getMessage(Long id) {
 		
 		return messageDao.getMessageByTrip(id);
 		
 	}
 	
-	public Message getMessageById(int id){
+	public Message getMessageById(Long id){
 		return messageDao.getMessageById(id);
 	}
 	
-	public void removeMessageByChatId(int id){
+	public void removeMessageByChatId(Long id){
 		messageDao.removeMessage(id);
 	}
