+++ src/main/java/com/packt/web/service/RoleServiceInterface.java
@@ -4,6 +4,6 @@ import com.packt.web.bean.Role;
 
 public interface RoleServiceInterface {
 
-	public Role getRole(int id);
+	public Role getRole(Long id);
 	
 }
